+++ ReleaseNotes.md
@@ -8,6 +8,7 @@ Release notes:
 * ImageN 0.9.1
 * Migrate to Spring Framework 6.2.12 and Jakarta Servlet 6.0.0
 * Update metrics to 4.2.37
+* Migrate to Apache HttpClient 5
 
 ## Release 2.4.0